LayoutTests/perf/sibling-renderer-On2-expected.txt

PASS timeRenderTreeCreation(5000) < 20 * timeRenderTreeCreation(500) is true
PASS successfullyParsed is true

TEST COMPLETE
